Address PR #49782 review comments (round 2)
- Move test-only SFPU wrappers (calculate_sqrt_custom, calculate_expm1_cw,
  calculate_mask_binary) out of sfpu_operations.h into a dedicated
  sfpu_test_helpers.h so the harness file only holds dispatch logic.
- Keep DOMAIN_MATHOPS in test_sfpu_unary.py in alphabetical order.
