+++ tt_metal/tt-llk/tests/python_tests/test_sfpu_unary.py
@@ -389,58 +389,58 @@ def test_eltwise_unary_sfpu_int(
 
 # Unary SFPU ops that require per-op input domains
 DOMAIN_MATHOPS = [
-    MathOperation.Erfinv,
-    MathOperation.Heaviside,
-    MathOperation.Softshrink,
-    MathOperation.Softsign,
-    MathOperation.Sigmoid,
-    MathOperation.Mish,
-    MathOperation.Selu,
-    MathOperation.I0,
-    MathOperation.EqualZero,
-    MathOperation.NotEqualZero,
-    MathOperation.LessThanZero,
-    MathOperation.GreaterThanZero,
-    MathOperation.LessThanEqualZero,
-    MathOperation.GreaterThanEqualZero,
-    MathOperation.Rdiv,
-    MathOperation.Clamp,
-    MathOperation.Hardtanh,
-    MathOperation.Lrelu,
-    MathOperation.Erf,
-    MathOperation.Erfc,
-    MathOperation.Expm1,
-    MathOperation.Cbrt,
-    MathOperation.I1,
-    MathOperation.Sign,
-    MathOperation.Signbit,
-    MathOperation.TanhDerivative,
-    MathOperation.Hardmish,
-    MathOperation.Lgamma,
-    MathOperation.Digamma,
-    MathOperation.Identity,
-    MathOperation.Prelu,
-    MathOperation.Rpow,
-    MathOperation.UnaryPower,
-    MathOperation.Fmod,
-    MathOperation.Remainder,
-    MathOperation.UnaryGt,
-    MathOperation.UnaryLt,
-    MathOperation.UnaryGe,
-    MathOperation.UnaryLe,
-    MathOperation.UnaryMax,
-    MathOperation.UnaryMin,
-    MathOperation.Polygamma,
-    MathOperation.Xielu,
-    MathOperation.Hardshrink,
-    MathOperation.Softplus,
-    MathOperation.SigmoidAppx,
-    MathOperation.SqrtCustom,
     MathOperation.Add1,
     MathOperation.CastFp32ToFp16a,
-    MathOperation.TanhDerivativeLut,
-    MathOperation.RsqrtCompat,
+    MathOperation.Cbrt,
+    MathOperation.Clamp,
+    MathOperation.Digamma,
+    MathOperation.EqualZero,
+    MathOperation.Erf,
+    MathOperation.Erfc,
+    MathOperation.Erfinv,
+    MathOperation.Expm1,
     MathOperation.Expm1Cw,
+    MathOperation.Fmod,
+    MathOperation.GreaterThanEqualZero,
+    MathOperation.GreaterThanZero,
+    MathOperation.Hardmish,
+    MathOperation.Hardshrink,
+    MathOperation.Hardtanh,
+    MathOperation.Heaviside,
+    MathOperation.I0,
+    MathOperation.I1,
+    MathOperation.Identity,
+    MathOperation.LessThanEqualZero,
+    MathOperation.LessThanZero,
+    MathOperation.Lgamma,
+    MathOperation.Lrelu,
+    MathOperation.Mish,
+    MathOperation.NotEqualZero,
+    MathOperation.Polygamma,
+    MathOperation.Prelu,
+    MathOperation.Rdiv,
+    MathOperation.Remainder,
+    MathOperation.Rpow,
+    MathOperation.RsqrtCompat,
+    MathOperation.Selu,
+    MathOperation.Sigmoid,
+    MathOperation.SigmoidAppx,
+    MathOperation.Sign,
+    MathOperation.Signbit,
+    MathOperation.Softplus,
+    MathOperation.Softshrink,
+    MathOperation.Softsign,
+    MathOperation.SqrtCustom,
+    MathOperation.TanhDerivative,
+    MathOperation.TanhDerivativeLut,
+    MathOperation.UnaryGe,
+    MathOperation.UnaryGt,
+    MathOperation.UnaryLe,
+    MathOperation.UnaryLt,
+    MathOperation.UnaryMax,
+    MathOperation.UnaryMin,
+    MathOperation.UnaryPower,
+    MathOperation.Xielu,
 ]
 
 # Per-op tolerance overrides for the domain suite. Ops whose kernel is a coarse
